# apps/payment-sandbox/docs/openapi/payment-sandbox.v1.json
{
  "components": {
    "schemas": {
      "BalanceProjectionLine": {
        "additionalProperties": false,
        "properties": {
          "account_type": {
            "enum": [
              "available",
              "reserved",
              "liquidable"
            ],
            "type": "string"
          },
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "currency": {
            "type": "string"
          },
          "merchant_id": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "merchant_id",
          "currency",
          "account_type",
          "amount",
          "updated_at"
        ],
        "type": "object"
      },
      "BalanceProjectionView": {
        "additionalProperties": false,
        "properties": {
          "balances": {
            "items": {
              "$ref": "#/components/schemas/BalanceProjectionLine"
            },
            "type": "array"
          },
          "count": {
            "format": "int64",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "balances",
          "count"
        ],
        "type": "object"
      },
      "CapturePaymentIntentRequest": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "description": "Optional capture amount override. Omit to capture the full remaining amount.",
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "idempotency_key": {
            "description": "Optional request idempotency key. If omitted, the `Idempotency-Key` header is used when present.",
            "type": "string"
          }
        },
        "type": "object"
      },
      "CapturePaymentIntentResponse": {
        "additionalProperties": false,
        "properties": {
          "charge": {
            "$ref": "#/components/schemas/ChargeView"
          },
          "payment_intent": {
            "$ref": "#/components/schemas/PaymentIntentView"
          }
        },
        "required": [
          "payment_intent",
          "charge"
        ],
        "type": "object"
      },
      "ChargeView": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "captured_amount": {
            "format": "int64",
            "type": "integer"
          },
          "captured_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "payment_attempt_id": {
            "type": "string"
          },
          "payment_intent_id": {
            "type": "string"
          },
          "refunded_amount": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "payment_intent_id",
          "amount",
          "captured_amount",
          "refunded_amount",
          "status",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "ChargeViewEnvelope": {
        "additionalProperties": false,
        "properties": {
          "charge": {
            "$ref": "#/components/schemas/ChargeView"
          }
        },
        "required": [
          "charge"
        ],
        "type": "object"
      },
      "ConfirmPaymentIntentRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "description": "Optional request idempotency key fallback.",
            "type": "string"
          },
          "payment_method_token": {
            "description": "Optional payment method token to drive the scenario.",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ConfirmPaymentIntentResponse": {
        "additionalProperties": false,
        "properties": {
          "charge": {
            "$ref": "#/components/schemas/ChargeView"
          },
          "payment_attempt": {
            "$ref": "#/components/schemas/PaymentAttemptView"
          },
          "payment_intent": {
            "$ref": "#/components/schemas/PaymentIntentView"
          }
        },
        "required": [
          "payment_intent",
          "payment_attempt"
        ],
        "type": "object"
      },
      "CreatePaymentIntentRequest": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "description": "Amount in the smallest currency unit.",
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "capture_method": {
            "default": "manual",
            "enum": [
              "manual",
              "automatic"
            ],
            "type": "string"
          },
          "currency": {
            "description": "ISO currency code.",
            "minLength": 1,
            "type": "string"
          },
          "customer_id": {
            "description": "Optional customer identifier.",
            "type": "string"
          },
          "idempotency_key": {
            "description": "Optional request idempotency key.",
            "type": "string"
          },
          "merchant_id": {
            "description": "Optional merchant identifier.",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency"
        ],
        "type": "object"
      },
      "CreatePaymentIntentResponse": {
        "additionalProperties": false,
        "properties": {
          "payment_intent": {
            "$ref": "#/components/schemas/PaymentIntentView"
          }
        },
        "required": [
          "payment_intent"
        ],
        "type": "object"
      },
      "FeeView": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "charge_id": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "charge_id",
          "amount",
          "currency",
          "created_at"
        ],
        "type": "object"
      },
      "PaymentAttemptView": {
        "additionalProperties": false,
        "properties": {
          "decline_code": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "payment_intent_id": {
            "type": "string"
          },
          "payment_method_token": {
            "type": "string"
          },
          "processor_reference": {
            "type": "string"
          },
          "requested_at": {
            "format": "date-time",
            "type": "string"
          },
          "responded_at": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "payment_intent_id",
          "status",
          "requested_at",
          "responded_at"
        ],
        "type": "object"
      },
      "PaymentAttemptViewEnvelope": {
        "additionalProperties": false,
        "properties": {
          "payment_attempt": {
            "$ref": "#/components/schemas/PaymentAttemptView"
          }
        },
        "required": [
          "payment_attempt"
        ],
        "type": "object"
      },
      "PaymentIntentView": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "capture_method": {
            "type": "string"
          },
          "charge_id": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "customer_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "idempotency_key": {
            "type": "string"
          },
          "latest_attempt_id": {
            "type": "string"
          },
          "merchant_id": {
            "type": "string"
          },
          "scenario": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "amount",
          "currency",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "PaymentIntentViewEnvelope": {
        "additionalProperties": false,
        "properties": {
          "payment_intent": {
            "$ref": "#/components/schemas/PaymentIntentView"
          }
        },
        "required": [
          "payment_intent"
        ],
        "type": "object"
      },
      "PaymentLifecycleView": {
        "additionalProperties": false,
        "properties": {
          "charge": {
            "$ref": "#/components/schemas/ChargeView"
          },
          "is_refundable": {
            "type": "boolean"
          },
          "latest_attempt": {
            "$ref": "#/components/schemas/PaymentAttemptView"
          },
          "payment_intent": {
            "$ref": "#/components/schemas/PaymentIntentView"
          },
          "refundable_amount": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "payment_intent",
          "status",
          "refundable_amount",
          "is_refundable"
        ],
        "type": "object"
      },
      "PaymentLifecycleViewEnvelope": {
        "additionalProperties": false,
        "properties": {
          "payment_lifecycle": {
            "$ref": "#/components/schemas/PaymentLifecycleView"
          }
        },
        "required": [
          "payment_lifecycle"
        ],
        "type": "object"
      },
      "RefundRequest": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "description": "Optional refund amount override. Omit to refund the remaining captured amount.",
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "charge_id": {
            "description": "Charge to refund.",
            "type": "string"
          },
          "idempotency_key": {
            "description": "Optional request idempotency key. If omitted, the `Idempotency-Key` header is used when present.",
            "type": "string"
          }
        },
        "required": [
          "charge_id"
        ],
        "type": "object"
      },
      "RefundResponse": {
        "additionalProperties": false,
        "properties": {
          "charge": {
            "$ref": "#/components/schemas/ChargeView"
          },
          "refund": {
            "$ref": "#/components/schemas/RefundView"
          }
        },
        "required": [
          "refund",
          "charge"
        ],
        "type": "object"
      },
      "RefundView": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "charge_id": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "payment_intent_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "charge_id",
          "payment_intent_id",
          "amount",
          "status",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "RefundViewEnvelope": {
        "additionalProperties": false,
        "properties": {
          "refund": {
            "$ref": "#/components/schemas/RefundView"
          }
        },
        "required": [
          "refund"
        ],
        "type": "object"
      },
      "SettlementBatchLine": {
        "additionalProperties": false,
        "properties": {
          "charge_count": {
            "format": "int64",
            "type": "integer"
          },
          "charge_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "charges": {
            "items": {
              "$ref": "#/components/schemas/SettlementChargeLine"
            },
            "type": "array"
          },
          "currency": {
            "type": "string"
          },
          "gross_amount": {
            "format": "int64",
            "type": "integer"
          },
          "merchant_id": {
            "type": "string"
          },
          "net_amount": {
            "format": "int64",
            "type": "integer"
          },
          "refund_count": {
            "format": "int64",
            "type": "integer"
          },
          "refunded_amount": {
            "format": "int64",
            "type": "integer"
          },
          "settlement_date": {
            "format": "date",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "ready"
            ],
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "merchant_id",
          "currency",
          "settlement_date",
          "status",
          "gross_amount",
          "refunded_amount",
          "net_amount",
          "charge_count",
          "refund_count",
          "charge_ids",
          "charges",
          "updated_at"
        ],
        "type": "object"
      },
      "SettlementChargeLine": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "captured_at": {
            "format": "date-time",
            "type": "string"
          },
          "charge_id": {
            "type": "string"
          },
          "payment_intent_id": {
            "type": "string"
          },
          "refunded_amount": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "charge_id",
          "payment_intent_id",
          "captured_at",
          "amount",
          "refunded_amount"
        ],
        "type": "object"
      },
      "SettlementProjectionView": {
        "additionalProperties": false,
        "properties": {
          "batches": {
            "items": {
              "$ref": "#/components/schemas/SettlementBatchLine"
            },
            "type": "array"
          },
          "count": {
            "format": "int64",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "batches",
          "count"
        ],
        "type": "object"
      },
      "TransactionReportEnvelope": {
        "additionalProperties": false,
        "properties": {
          "transactions_report": {
            "$ref": "#/components/schemas/TransactionReportView"
          }
        },
        "required": [
          "transactions_report"
        ],
        "type": "object"
      },
      "TransactionReportLine": {
        "additionalProperties": false,
        "properties": {
          "charge": {
            "$ref": "#/components/schemas/ChargeView"
          },
          "fees": {
            "items": {
              "$ref": "#/components/schemas/FeeView"
            },
            "type": "array"
          },
          "latest_attempt": {
            "$ref": "#/components/schemas/PaymentAttemptView"
          },
          "payment_intent": {
            "$ref": "#/components/schemas/PaymentIntentView"
          },
          "refunds": {
            "items": {
              "$ref": "#/components/schemas/RefundView"
            },
            "type": "array"
          }
        },
        "required": [
          "payment_intent"
        ],
        "type": "object"
      },
      "TransactionReportView": {
        "additionalProperties": false,
        "properties": {
          "balance_projection": {
            "$ref": "#/components/schemas/BalanceProjectionView"
          },
          "count": {
            "format": "int64",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "type": "string"
          },
          "settlement_projection": {
            "$ref": "#/components/schemas/SettlementProjectionView"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionReportLine"
            },
            "type": "array"
          }
        },
        "required": [
          "generated_at",
          "transactions",
          "balance_projection",
          "settlement_projection",
          "count"
        ],
        "type": "object"
      },
      "TransactionSnapshotEnvelope": {
        "additionalProperties": false,
        "properties": {
          "transactions_snapshot": {
            "$ref": "#/components/schemas/TransactionReportView"
          }
        },
        "required": [
          "transactions_snapshot"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Contract for integrating workflow consumers with the payment sandbox.",
    "license": {
      "identifier": "UNLICENSED",
      "name": "UNLICENSED"
    },
    "title": "Payment Sandbox API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/charges/{id}": {
      "get": {
        "operationId": "getChargeV1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChargeViewEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "summary": "Fetch charge"
      }
    },
    "/payment_attempts/{id}": {
      "get": {
        "operationId": "getPaymentAttemptV1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAttemptViewEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "summary": "Fetch payment attempt"
      }
    },
    "/payment_intents": {
      "post": {
        "operationId": "createPaymentIntentV1",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "amount": 1000,
                "capture_method": "manual",
                "currency": "usd",
                "customer_id": "cus_123",
                "idempotency_key": "create-123",
                "merchant_id": "merchant_123"
              },
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentIntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePaymentIntentResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "summary": "Create a payment intent"
      }
    },
    "/payment_intents/{id}": {
      "get": {
        "operationId": "getPaymentIntentV1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIntentViewEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "summary": "Fetch payment status"
      }
    },
    "/payment_intents/{id}/capture": {
      "post": {
        "operationId": "capturePaymentIntentV1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "amount": 1000,
                "idempotency_key": "capture-123"
              },
              "schema": {
                "$ref": "#/components/schemas/CapturePaymentIntentRequest"
              }
            }
          },
          "description": "Optional capture overrides. Omit the body to capture the full remaining amount.",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapturePaymentIntentResponse"
                }
              }
            },
            "description": "Captured"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Capture a payment intent"
      }
    },
    "/payment_intents/{id}/confirm": {
      "post": {
        "operationId": "confirmPaymentIntentV1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional sandbox scenario override for deterministic behavior.",
            "in": "header",
            "name": "X-Sandbox-Scenario",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional idempotency key. The body field is also accepted as a fallback.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "idempotency_key": "confirm-123",
                "payment_method_token": "pm_card_visa"
              },
              "schema": {
                "$ref": "#/components/schemas/ConfirmPaymentIntentRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmPaymentIntentResponse"
                }
              }
            },
            "description": "Confirmed"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Confirm a payment intent"
      }
    },
    "/payment_intents/{id}/lifecycle": {
      "get": {
        "operationId": "getPaymentLifecycleV1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLifecycleViewEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "summary": "Fetch payment lifecycle"
      }
    },
    "/refunds": {
      "post": {
        "operationId": "createRefundV1",
        "parameters": [
          {
            "description": "Optional request idempotency key. The body field is also accepted as a fallback.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "amount": 1000,
                "charge_id": "ch_123",
                "idempotency_key": "refund-123"
              },
              "schema": {
                "$ref": "#/components/schemas/RefundRequest"
              }
            }
          },
          "description": "Refund body. `charge_id` is required; `amount` and `idempotency_key` are optional.",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "summary": "Create a refund"
      }
    },
    "/refunds/{id}": {
      "get": {
        "operationId": "getRefundV1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundViewEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "summary": "Fetch refund"
      }
    },
    "/reports/transactions": {
      "get": {
        "operationId": "getTransactionReportV1",
        "parameters": [
          {
            "description": "Optional export mode. `snapshot` returns the reconciliation export envelope.",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "enum": [
                "report",
                "snapshot"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TransactionReportEnvelope"
                    },
                    {
                      "$ref": "#/components/schemas/TransactionSnapshotEnvelope"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Fetch transaction report"
      }
    }
  },
  "security": [],
  "servers": [
    {
      "url": "/v1"
    }
  ]
}