# apps/payment-sandbox/internal/adapters/inbound/http/server.go
package httpadapter

import (
	"net/http"
	"strings"

	docsadapter "payment-sandbox/internal/adapters/inbound/http/docs"
	"payment-sandbox/internal/domain"
	"payment-sandbox/internal/sandbox"
)

type Server struct {
	svc *sandbox.Service
	mux *http.ServeMux
}

func New(svc *sandbox.Service, docsRoot ...string) http.Handler {
	s := &Server{svc: svc, mux: http.NewServeMux()}
	s.routes()
	root := ""
	if len(docsRoot) > 0 {
		root = docsRoot[0]
	}
	s.mux.Handle("GET /openapi/", docsadapter.New(root))
	return s.mux
}

func (s *Server) routes() {
	s.mux.HandleFunc("GET /health", s.handleHealth)
	s.mux.HandleFunc("GET /v1/payment_intents/{id}", s.handleGetPaymentIntent)
	s.mux.HandleFunc("GET /v1/payment_attempts/{id}", s.handleGetPaymentAttempt)
	s.mux.HandleFunc("GET /v1/charges/{id}", s.handleGetCharge)
	s.mux.HandleFunc("GET /v1/refunds/{id}", s.handleGetRefund)
	s.mux.HandleFunc("GET /v1/payment_intents/{id}/lifecycle", s.handleGetPaymentLifecycle)
	s.mux.HandleFunc("GET /v1/reports/transactions", s.handleGetTransactionReport)
	s.mux.HandleFunc("POST /v1/payment_intents", s.handleCreatePaymentIntent)
	s.mux.HandleFunc("POST /v1/payment_intents/{id}/confirm", s.handleConfirmPaymentIntent)
	s.mux.HandleFunc("POST /v1/payment_intents/{id}/capture", s.handleCapturePaymentIntent)
	s.mux.HandleFunc("POST /v1/refunds", s.handleCreateRefund)
}

func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
	WriteJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

func (s *Server) handleCreatePaymentIntent(w http.ResponseWriter, r *http.Request) {
	var req sandbox.CreatePaymentIntentRequest
	payload, err := ReadJSON(r, &req)
	if err != nil {
		writeError(w, err)
		return
	}
	idem := RequestIdempotencyKey(r, req.IdempotencyKey)
	result, err := s.svc.CreatePaymentIntent(req, idem, fingerprint(payload))
	if err != nil {
		writeError(w, err)
		return
	}
	WriteJSON(w, http.StatusCreated, sandbox.CreatePaymentIntentResponse{PaymentIntent: result})
}

func (s *Server) handleGetPaymentIntent(w http.ResponseWriter, r *http.Request) {
	intent, err := s.svc.GetPaymentIntent(r.PathValue("id"))
	if err != nil {
		writeError(w, err)
		return
	}
	WriteJSON(w, http.StatusOK, map[string]any{"payment_intent": intent})
}

func (s *Server) handleGetPaymentAttempt(w http.ResponseWriter, r *http.Request) {
	attempt, err := s.svc.GetPaymentAttempt(r.PathValue("id"))
	if err != nil {
		writeError(w, err)
		return
	}
	WriteJSON(w, http.StatusOK, map[string]any{"payment_attempt": attempt})
}

func (s *Server) handleGetCharge(w http.ResponseWriter, r *http.Request) {
	charge, err := s.svc.GetCharge(r.PathValue("id"))
	if err != nil {
		writeError(w, err)
		return
	}
	WriteJSON(w, http.StatusOK, map[string]any{"charge": charge})
}

func (s *Server) handleGetRefund(w http.ResponseWriter, r *http.Request) {
	refund, err := s.svc.GetRefund(r.PathValue("id"))
	if err != nil {
		writeError(w, err)
		return
	}
	WriteJSON(w, http.StatusOK, map[string]any{"refund": refund})
}

func (s *Server) handleGetPaymentLifecycle(w http.ResponseWriter, r *http.Request) {
	lifecycle, err := s.svc.GetPaymentLifecycle(r.PathValue("id"))
	if err != nil {
		writeError(w, err)
		return
	}
	WriteJSON(w, http.StatusOK, map[string]any{"payment_lifecycle": lifecycle})
}

func (s *Server) handleGetTransactionReport(w http.ResponseWriter, r *http.Request) {
	mode, err := transactionReportModeFromRequest(r)
	if err != nil {
		writeError(w, err)
		return
	}
	report, err := s.svc.GetTransactionReport(mode)
	if err != nil {
		writeError(w, err)
		return
	}
	response := map[string]any{"transactions_report": report}
	if mode == sandbox.TransactionReportModeSnapshot {
		response = map[string]any{"transactions_snapshot": report}
	}
	WriteJSON(w, http.StatusOK, response)
}

func transactionReportModeFromRequest(r *http.Request) (sandbox.TransactionReportMode, error) {
	view := strings.TrimSpace(r.URL.Query().Get("view"))
	switch view {
	case "", "report":
		return sandbox.TransactionReportModeCurrent, nil
	case "snapshot":
		return sandbox.TransactionReportModeSnapshot, nil
	default:
		return "", domain.NewError(http.StatusBadRequest, "invalid_request", "invalid view query parameter")
	}
}

func (s *Server) handleConfirmPaymentIntent(w http.ResponseWriter, r *http.Request) {
	var req sandbox.ConfirmPaymentIntentRequest
	payload, err := ReadJSON(r, &req)
	if err != nil {
		writeError(w, err)
		return
	}
	intentID := r.PathValue("id")
	scenarioHeader := strings.TrimSpace(r.Header.Get("X-Sandbox-Scenario"))
	idem := RequestIdempotencyKey(r, req.IdempotencyKey)
	fingerprint := fingerprintString(string(payload) + "|scenario=" + scenarioHeader + "|token=" + req.PaymentMethodToken)
	result, err := s.svc.ConfirmPaymentIntent(intentID, req, scenarioHeader, idem, fingerprint)
	if err != nil {
		writeError(w, err)
		return
	}
	WriteJSON(w, http.StatusOK, result)
}

func (s *Server) handleCapturePaymentIntent(w http.ResponseWriter, r *http.Request) {
	var req sandbox.CapturePaymentIntentRequest
	payload, err := ReadJSON(r, &req)
	if err != nil {
		writeError(w, err)
		return
	}
	intentID := r.PathValue("id")
	idem := RequestIdempotencyKey(r, req.IdempotencyKey)
	fingerprint := fingerprintString(string(payload) + "|intent_id=" + intentID + "|idempotency=" + idem)
	result, err := s.svc.CapturePaymentIntent(intentID, req, fingerprint)
	if err != nil {
		writeError(w, err)
		return
	}
	WriteJSON(w, http.StatusOK, result)
}

func (s *Server) handleCreateRefund(w http.ResponseWriter, r *http.Request) {
	var req sandbox.RefundRequest
	payload, err := ReadJSON(r, &req)
	if err != nil {
		writeError(w, err)
		return
	}
	idem := RequestIdempotencyKey(r, req.IdempotencyKey)
	result, err := s.svc.CreateRefund(req, idem, fingerprint(payload))
	if err != nil {
		writeError(w, err)
		return
	}
	WriteJSON(w, http.StatusCreated, result)
}